fortitwin_mvp/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
google-generativeai==0.7.2
chromadb==0.5.5
numpy==1.26.4
tqdm==4.66.5
websockets==12.0
